Покрыл код комментариями

+++ src/main/java/me/func/parametricfunction/factory/TextAreaFactory.java
@@ -16,6 +16,7 @@ public class TextAreaFactory {
     public TextArea[] createNumberTextAreas(int xProperty, int yProperty, int zProperty, boolean isNormal) {
         TextArea[] numbers = new TextArea[TEXT_AREA_OFFSET];
 
+        // Создание текстовых областей на основе заданных свойств и флага
         for (int i = -TEXT_AREA_HALF_LENGTH_OFFSET; i < TEXT_AREA_HALF_LENGTH_OFFSET; i++) {
             int index = i + TEXT_AREA_HALF_LENGTH_OFFSET;
             numbers[index] = initializeTextArea(i, xProperty, yProperty, zProperty, isNormal);
@@ -32,16 +33,26 @@ public class TextAreaFactory {
         return textArea;
     }
 
-    private void setupTextAreaProperties(TextArea textArea, int value, int xProperty, int yProperty, int zProperty, boolean isNormal) {
+    private void setupTextAreaProperties(
+            TextArea textArea,
+            int value,
+            int xProperty,
+            int yProperty,
+            int zProperty,
+            boolean isNormal
+    ) {
+        // Установка позиции и размера текстовой области на основе переданных свойств и значения
         textArea.translateXProperty().set(-value * TEXT_AREA_TRANSLATE_FACTOR * xProperty);
         textArea.translateYProperty().set(value * TEXT_AREA_TRANSLATE_FACTOR * yProperty);
 
+        // Если это нормальная текстовая область, производим дополнительные настройки
         if (isNormal) {
-            textArea.translateYProperty().set(value * TEXT_AREA_TRANSLATE_FACTOR * yProperty - (double) TEXT_AREA_TRANSLATE_FACTOR / 2);
+            textArea.translateYProperty().set(value * TEXT_AREA_TRANSLATE_FACTOR * yProperty - TEXT_AREA_TRANSLATE_FACTOR / 2);
             textArea.setRotationAxis(Rotate.X_AXIS);
             textArea.setRotate(270);
         }
 
+        // Настройка размеров, шрифта и других свойств текстовой области
         textArea.translateZProperty().set(value * TEXT_AREA_TRANSLATE_FACTOR * zProperty);
         textArea.setMaxWidth(TEXT_AREA_TRANSLATE_FACTOR);
         textArea.setMaxHeight(TEXT_AREA_TRANSLATE_FACTOR);
